requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
openai>=1.0.0
redis==5.0.1
cryptography==41.0.7
